+++ backend/src/main/java/se/example2/softhouse/core/ExamQuestion.java
@@ -9,37 +9,37 @@ import org.hibernate.validator.constraints.NotEmpty;
 public class ExamQuestion {
 
     @JsonProperty
-    private Long id;
+    private int id;
 
     @JsonProperty
     @NotEmpty
-    private Long examId;
+    private int examId;
 
     @JsonProperty
     @NotEmpty
-    private Long questionId;
+    private int questionId;
 
-    public Long getQuestionId() {
+    public int getQuestionId() {
         return questionId;
     }
 
-    public void setQuestionId(Long questionId) {
+    public void setQuestionId(int questionId) {
         this.questionId = questionId;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getExamId() {
+    public int getExamId() {
         return examId;
     }
 
-    public void setExamId(Long examId) {
+    public void setExamId(int examId) {
         this.examId = examId;
     }
